+++ task8-p2p-chat/.gitignore
@@ -4,3 +4,6 @@ target/
 build/
 out/
 log/
+*.lck
+*.vux
+*.bak*
