+++ task8-p2p-chat/p2chat-commons/src/main/kotlin/ru/mit/spbau/sd/chat/commons/net/simpleProtocol.kt
@@ -1,5 +1,6 @@
-package ru.mit.spbau.sd.chat.commons
+package ru.mit.spbau.sd.chat.commons.net
 
+import ru.mit.spbau.sd.chat.commons.intSizeInBytes
 import ru.mit.spbau.sd.chat.commons.net.*
 import java.nio.ByteBuffer
