+++ task8-p2p-chat/p2chat-commons/src/main/proto/message.proto
@@ -57,7 +57,7 @@ message PeerToPeerMsg {
     enum Type {
         TEXT_MESSAGE = 0;
         MY_INFO_CHANGED = 1;
-
+        I_GO_OFFLINE = 2;
     }
     ChatMessage message = 2;
 }
