+++ task8-p2p-chat/p2chat-commons/src/test/kotlin/ru/mit/spbau/sd/chat/commons/WritingStateTest.kt
@@ -4,6 +4,7 @@ import org.junit.Assert
 import org.junit.Test
 import ru.mit.spbau.sd.chat.commons.net.GatheringWriting
 import ru.mit.spbau.sd.chat.commons.net.WritingState
+import ru.mit.spbau.sd.chat.commons.net.createStartWritingState
 import java.nio.ByteBuffer
