+++ task8-p2p-chat/p2chat-server/src/main/kotlin/ru/mit/spbau/sd/chat/server/server.kt
@@ -1,10 +1,11 @@
 package ru.mit.spbau.sd.chat.server
 
+import ru.mit.spbau.sd.chat.commons.ConcurrentUsersMap
 import ru.mit.spbau.sd.chat.server.net.ChatModelPeerMsgProcessor
 import ru.mit.spbau.sd.chat.server.net.ChatServer
 
 fun main(args: Array<String>) {
-    val chatModel = ConcurrentChatModel()
+    val chatModel = ConcurrentUsersMap()
     val peerEventProcessor = ChatModelPeerMsgProcessor(chatModel)
     val chatServer = ChatServer(peerEventProcessor)
